requirements.txt
discord
pymongo
dnspython
schedule
numpy
dnspython
coincurve
pysha3
secrets
